+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 ##[0.2.4] - [2016-02-06]
 ###Added
-- *set!*, *let**
+- *set!*, *named let*, *let**, *letrec(let)*, and *letrec*(let*)* -
+  see test files for samples.
 - more tests
 
 ###Fixed